+++ src/lib/djerba/plugins/virus/test/plugin_test.py
@@ -1,24 +1,20 @@
 #! /usr/bin/env python3
 
 """
-Test of the WGTS small plugin
-AUTHOR: Felix Beaudry
+Test for VIRUSBreakend plugin.
 """
 
 import os
 import unittest
 import tempfile
 import string
-
 from djerba.util.validator import path_validator
 from djerba.plugins.plugin_tester import PluginTester
-import djerba.plugins.wgts.cnv_purple.plugin as cnv
-from djerba.core.workspace import workspace
 from djerba.util.environment import directory_finder
 
-class TestPurplePlugin(PluginTester):
+class testVirus(PluginTester):
 
-    WGTS_INI_NAME = 'cnv.wgts.ini'
+    INI_NAME = 'virus.ini'
 
     def setUp(self):
         self.path_validator = path_validator()
@@ -26,30 +22,25 @@ class TestPurplePlugin(PluginTester):
         self.tmp = tempfile.TemporaryDirectory(prefix='djerba_')
         self.tmp_dir = self.tmp.name
         self.sup_dir = directory_finder().get_test_dir()
+        self.json = os.path.join(self.sup_dir, "plugins/virus/virus.json")
 
-    def testWGTScnv(self):
+    def test(self):
         test_source_dir = os.path.realpath(os.path.dirname(__file__))
-        with open(os.path.join(test_source_dir, self.WGTS_INI_NAME)) as in_file:
+        with open(os.path.join(test_source_dir, self.INI_NAME)) as in_file:
             template_str = in_file.read()
         template = string.Template(template_str)
         ini_str = template.substitute({'DJERBA_TEST_DATA': self.sup_dir})
         input_dir = os.path.join(self.get_tmp_dir(), 'input')
         os.mkdir(input_dir)
-        with open(os.path.join(input_dir, self.WGTS_INI_NAME), 'w') as ini_file:
+        with open(os.path.join(input_dir, self.INI_NAME), 'w') as ini_file:
             ini_file.write(ini_str)
-        json_location = os.path.join(self.sup_dir ,"plugins/cnv-purple/report_json/cnv.purple.json")
+
         params = {
-            self.INI: self.WGTS_INI_NAME,
-            self.JSON: json_location,
-            self.MD5: '5c15283f5b0ee48201a980ac5ef721dc'
+            self.INI: self.INI_NAME,
+            self.JSON: self.json,
+            self.MD5: '8188c77f8f36a222a61b0a76c2ddecc0'
         }
         self.run_basic_test(input_dir, params)
 
-    def redact_json_data(self, data):
-        """replaces empty method from testing.tools"""
-        for key in ['cnv plot']:
-            del data['results'][key]
-        return data 
-    
 if __name__ == '__main__':
     unittest.main()