+++ src/test/core/test_core.py
@@ -36,8 +36,9 @@ class TestCore(TestBase):
     LOREM_FILENAME = 'lorem.txt'
     SIMPLE_REPORT_JSON = 'simple_report_expected.json'
     SIMPLE_REPORT_UPDATE_JSON = 'simple_report_for_update.json'
+    SIMPLE_REPORT_UPDATE_FAILED_JSON = 'simple_report_for_update_failed.json'
     SIMPLE_CONFIG_MD5 = '04b749b3ec489ed9c06c1a06eb2dc886'
-    SIMPLE_REPORT_MD5 = 'ab049488c58758e26b0ad1c480c28c99'
+    SIMPLE_REPORT_MD5 = 'cfa53b636c7e8ae0f78fff698c4f76b7'
 
     class mock_args:
         """Use instead of argparse to store params for testing"""
@@ -182,9 +183,9 @@ class TestConfigValidation(TestCore):
         self.assertTrue(plugin.check_attributes_known(attributes))
         config.set('demo1', 'attributes', 'clinical,awesome')
         attributes = plugin.get_config_wrapper(config).get_my_attributes()
-        with self.assertLogs('djerba.core.configure', level=logging.WARNING) as log_context:
+        with self.assertLogs('djerba:demo1', level=logging.WARNING) as log_context:
             self.assertFalse(plugin.check_attributes_known(attributes))
-        msg = "WARNING:djerba.core.configure:Unknown attribute 'awesome' in config"
+        msg = "WARNING:djerba:demo1:Unknown attribute 'awesome' in config"
         self.assertIn(msg, log_context.output)
 
     def test_simple(self):
@@ -192,9 +193,9 @@ class TestConfigValidation(TestCore):
         config = self.read_demo1_config(plugin)
         # test a simple plugin
         self.assertTrue(plugin.validate_minimal_config(config))
-        with self.assertLogs('djerba.core.configure', level=logging.DEBUG) as log_context:
+        with self.assertLogs('djerba:demo1', level=logging.DEBUG) as log_context:
             self.assertTrue(plugin.validate_full_config(config))
-        msg = 'DEBUG:djerba.core.configure:'+\
+        msg = 'DEBUG:djerba:demo1:'+\
             '8 expected INI param(s) found for component demo1'
         self.assertIn(msg, log_context.output)
 
@@ -243,9 +244,9 @@ class TestConfigValidation(TestCore):
         # now give foo a config value
         config.set('demo1', 'foo', 'snark')
         self.assertTrue(plugin.validate_minimal_config(config))
-        with self.assertLogs('djerba.core.configure', level=logging.DEBUG) as log_context:
+        with self.assertLogs('djerba:demo1', level=logging.DEBUG) as log_context:
             self.assertTrue(plugin.validate_full_config(config))
-        msg = 'DEBUG:djerba.core.configure:'+\
+        msg = 'DEBUG:djerba:demo1:'+\
             '9 expected INI param(s) found for component demo1'
         self.assertIn(msg, log_context.output)
         # test setting all requirements
@@ -430,7 +431,8 @@ class TestHtmlCache(TestCore):
 
     def test_encode_decode(self):
         # test an encoding/decoding round trip
-        string_to_encode = "Hello, world!"
+        # including non-Latin characters (Greek alpha, beta, gamma, delta)
+        string_to_encode = "Hello, world! \u03b1\u03b2\u03b3\u03b4"
         cache = html_cache(log_level=logging.ERROR)
         encoded = cache.encode_to_base64(string_to_encode)
         decoded_string = cache.decode_from_base64(encoded)
@@ -620,6 +622,26 @@ class TestMainScript(TestCore):
         self.assertEqual(result.returncode, 0)
         self.assertSimpleReport(json_path, html)
 
+    def test_setup_cli(self):
+        mode = 'setup'
+        ini_path = os.path.join(self.tmp_dir, 'config.ini')
+        html = os.path.join(self.tmp_dir, 'placeholder_report.clinical.html')
+        cmd = [
+            'djerba.py', mode,
+            '--assay', 'wgts',
+            '--ini', ini_path,
+            '--compact'
+        ]
+        result = subprocess_runner().run(cmd)
+        self.assertEqual(result.returncode, 0)
+        self.assertEqual(self.getMD5(ini_path), 'a211144356b5ec200e1c31ecd3128b45')
+        os.remove(ini_path)
+        prepop_path = os.path.join(self.test_source_dir, 'prepop.ini')
+        cmd.extend(['--pre-populate', prepop_path])
+        result = subprocess_runner().run(cmd)
+        self.assertEqual(result.returncode, 0)
+        self.assertEqual(self.getMD5(ini_path), '2387e66d783b1deb0fe5361e7770ec7a')
+
     def test_update_cli_with_ini(self):
         mode = 'update'
         work_dir = self.tmp_dir
@@ -647,7 +669,7 @@ class TestMainScript(TestCore):
         html_path = os.path.join(self.tmp_dir, 'placeholder_report.clinical.html')
         with open(html_path) as html_file:
             html_string = html_file.read()
-        self.assert_report_MD5(html_string, '5bc52ffc10821f166fed7b3055cc8bad')
+        self.assert_report_MD5(html_string, 'a262bf44dc2d759f165bbe817ec16d22')
         pdf_path = os.path.join(self.tmp_dir, 'placeholder_report.clinical.pdf')
         self.assertTrue(os.path.isfile(pdf_path))
         updated_path = os.path.join(self.tmp_dir, 'simple_report_for_update.updated.json')
@@ -660,23 +682,38 @@ class TestMainScript(TestCore):
         summary_path = os.path.join(self.test_source_dir, 'alternate_summary.txt')
         # run djerba.py and check the results
         json_path = os.path.join(self.test_source_dir, self.SIMPLE_REPORT_UPDATE_JSON)
-        cmd = [
+        cmd_base = [
             'djerba.py', mode,
             '--work-dir', work_dir,
             '--summary', summary_path,
-            '--json', json_path,
             '--out-dir', self.tmp_dir,
             '--pdf'
         ]
+        cmd = cmd_base + ['--json', json_path]
         result = subprocess_runner().run(cmd)
         self.assertEqual(result.returncode, 0)
         html_path = os.path.join(self.tmp_dir, 'placeholder_report.clinical.html')
         with open(html_path) as html_file:
             html_string = html_file.read()
-        self.assert_report_MD5(html_string, '285adea0d50933a5da00c6f0452ba045')
+        self.assert_report_MD5(html_string, '781d477894d5a6e269cb68535a82ca89')
         pdf_path = os.path.join(self.tmp_dir, 'placeholder_report.clinical.pdf')
         self.assertTrue(os.path.isfile(pdf_path))
         updated_path = os.path.join(self.tmp_dir, 'simple_report_for_update.updated.json')
+        self.assertTrue(os.path.isfile(updated_path))
+        # test again with a failed report
+        for output in [html_path, pdf_path, updated_path]:
+            os.remove(output)
+        json_path = os.path.join(self.test_source_dir, self.SIMPLE_REPORT_UPDATE_FAILED_JSON)
+        cmd_failed = cmd_base + ['--json', json_path]
+        result = subprocess_runner().run(cmd_failed)
+        self.assertEqual(result.returncode, 0)
+        html_path = os.path.join(self.tmp_dir, 'placeholder_report.clinical.html')
+        with open(html_path) as html_file:
+            html_string = html_file.read()
+        self.assert_report_MD5(html_string, '093ca0030bcb2a69d9ac4d784a19b147')
+        pdf_path = os.path.join(self.tmp_dir, 'placeholder_report.clinical.pdf')
+        self.assertTrue(os.path.isfile(pdf_path))
+        updated_path = os.path.join(self.tmp_dir, 'simple_report_for_update_failed.updated.json')
         self.assertTrue(os.path.isfile(updated_path))