+++ CHANGELOG.md
@@ -1,11 +1,42 @@
 # CHANGELOG
 
+## Unreleased: v1.8.0
+- GCGI-1473: Remove R code from wgts.snv_indel and wgts.cnv_purple plugins
+
+## 1.7.9: 2025-01-22
+- GCGI-1461: Fix output paths in calls to get_logger
+- GCGI-1462: Extend GSICAPBENCH report to TAR/PWGS
+- GCGI-1481: Fix raw coverage auto-population to exclude normal samples before selection in TAR assay
+- GCGI-1478: More informative logger name for plugin/helper/merger components
+- GCGI-1479: New `--pre-populate` option in `djerba.py` setup mode
+- GCGI-1413: Remove failed report plugin and allow summary plugin to handle failed reports
+- GCGI-1482: Updated total genome segment length constant in percent genome altered calculation
+- GCGI-1480: Updated CGI manager name and email
+- GCGI-1490: Remove input paths from pwgs.analysis results
+- GCGI-1492: Remove the Sequenza CNV plugin
+
+## 1.7.8: 2024-12-12
+- GCGI-1464: Standalone script to diff two Djerba JSON reports
+- GCGI-1454: Added OncoKB definitions to WGTS40X and WGS40X assays
+- GCGI-1469: Prevent provenance_helper from overwriting manually provided tumour and normal IDs
+- GCGI-1472: Add all patient IDs to case_overview (not just the initial ID)
+- GCGI-1326: Generate whizbam links to view fusions in split-view
+
+## 1.7.7: 2024-11-11
+- GCGI-1424: Fix rendering of non-Latin characters
+- GCGI-1437: Remove unnecessary checks in OncoKB cache
+- GCGI-981: Create virus (research) plugin
+- GCGI-1460: Make study a required parameter for PWGS
+- GCGI-1459: Update report date for PWGS assay from yyyy/mm/dd to date report was generated
+- GCGI-1458: Change geneticist sign-off date from date report was generated to yyyy-mm-dd
+- GCGI-1465: Fix for overzealous date format check from implementation of GCGI-1458
+
 ## 1.7.6: 2024-10-22
 - GCGI-1156: Merge long-running documentation branch to main
 - GCGI-1452: Update disclaimer with somatic mutation text and laboratory disclaimer
 - GCGI-1453: Update pWGS version number
 - GCGI-1456: Fix rounding of estimated tumour fraction in tar.sample
-- Increase csv field limit in plugins.fusions.tools.py to handle large fields in mavis_summary.txt 
+- Increase csv field limit in plugins.fusions.tools.py to handle large fields in mavis_summary.txt
 
 ## 1.7.5: 2024-10-02
 - GCGI-1378: Improve benchmarking to write an HTML summary page; update tests